+++ proxy/CoreUtils.cc
@@ -1154,7 +1154,7 @@ print_netstate(NetState * n)
   // These might not be 64-bit correct. /leif
   printf("      enabled: %d\n", n->enabled);
   printf("      op: %d  _cont: 0x%p\n", n->vio.op, n->vio._cont);
-  printf("      nbytes: %d  done: %d\n", n->vio.nbytes, n->vio.ndone);
+  printf("      nbytes: %d  done: %d\n", (int)n->vio.nbytes, (int)n->vio.ndone);
   printf("      vc_server: 0x%p   mutex: 0x%p\n\n", n->vio.vc_server, n->vio.mutex.m_ptr);
 }
